+++ src/flask_rq/_make.py
@@ -53,14 +53,9 @@ def make_queues(app: Flask | Quart) -> dict[str, Queue]:
     job_class = make_job_class(app)
     queues: dict[str, Queue] = {}
 
-    if "default" not in config.get("queues", []):
-        queues["default"] = Queue(
-            "default", default_conn, is_async=is_async, job_class=job_class
-        )
-
     # All defined connections have been created, now create each queue with the
     # connection it references.
-    for name in config.get("queues", []):
+    for name in config.get("queues", ["default"]):
         if name in connections:
             conn = connections[name]
         elif name in conn_refs: